+++ source/Internal/Container/Event/ConfigurationChangedEventSubscriber.php
@@ -9,18 +9,28 @@ declare(strict_types=1);
 
 namespace OxidEsales\EshopCommunity\Internal\Container\Event;
 
-use OxidEsales\EshopCommunity\Internal\Container\ContainerFactory;
 use OxidEsales\EshopCommunity\Internal\Framework\DIContainer\Event\ProjectYamlChangedEvent;
+use OxidEsales\EshopCommunity\Internal\Framework\DIContainer\Service\ContainerCache;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 class ConfigurationChangedEventSubscriber implements EventSubscriberInterface
 {
     /**
+     * @var ContainerCache
+     */
+    private $containerCache;
+
+    public function __construct(ContainerCache $containerCache)
+    {
+        $this->containerCache = $containerCache;
+    }
+
+    /**
      * @param ProjectYamlChangedEvent $event
      */
     public function resetContainer(ProjectYamlChangedEvent $event)
     {
-        ContainerFactory::resetContainer();
+        $this->containerCache->invalidate();
     }
 
     /**